Move BasicMath to expression and Delete class Memory
